refactor: clean up code and expand test coverage
- Remove unnecessary Default impls, use derive macro instead
- Remove unnecessary cast (u128 -> u128)
- Remove useless TryFrom impl for NetworkSize (always returns Ok)
- Collapse nested if statements (clippy)
- Add 11 new unit tests for fmt module functions:
  - test_short_peer_id, test_auto_scroll_offset, test_scroll_title
  - test_peer_display_name (local/received/fallback)
  - test_gen_msg_id (uniqueness, format)
  - test_format_system_time, test_format_latency (ms/s/none)

+++ src/fmt.rs
@@ -32,7 +32,7 @@ pub fn gen_msg_id() -> String {
     let now_ns = SystemTime::now()
         .duration_since(SystemTime::UNIX_EPOCH)
         .unwrap_or_default()
-        .as_nanos() as u128;
+        .as_nanos();
     let c = COUNTER.fetch_add(1, Ordering::Relaxed) as u128;
     format!("{:x}{:x}", now_ns, c)
 }
